requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
dnspython==1.16.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.1
#opencv-python==4.1.1.26
Pillow==6.1.0
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
reportlab==3.5.28
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3